+++ test/console.js
@@ -5,13 +5,13 @@
 const assert = require('assert');
 const util = require('util');
 
-const intel = require('../');
+const rufus = require('../');
 const consoleUtil = require('./util/console');
 
-var spy = new intel.handlers.Null();
+var spy = new rufus.handlers.Null();
 spy.handle = function(record) {
   this._lastRecord = record;
-  return intel.handlers.Null.prototype.handle.call(this, record);
+  return rufus.handlers.Null.prototype.handle.call(this, record);
 };
 
 var prevLog;
@@ -24,12 +24,12 @@ module.exports = {
 
   'console': {
     'before': function() {
-      intel.addHandler(spy);
+      rufus.addHandler(spy);
       prevLog = console.log;
       console.log = mockLog;
       // not passing root means this file becomes root.
       // which means dirname.basename, or test.console
-      intel.console();
+      rufus.console();
     },
     'can inject into global scope': function() {
       console.warn('test');
@@ -44,7 +44,7 @@ module.exports = {
       assert.equal(spy._lastRecord.name, 'test.util.console');
     },
     'can ignore paths': function() {
-      intel.console({ ignore: ['test.util'] });
+      rufus.console({ ignore: ['test.util'] });
 
       console.log('quux');
       assert.equal(spy._lastRecord.message, 'quux');
@@ -53,7 +53,7 @@ module.exports = {
       assert.notEqual(spy._lastRecord.message, 'baz');
       assert.equal(lastMock[0], 'baz');
 
-      intel.console();
+      rufus.console();
     },
     'overrides console.dir()': function() {
       var obj = { foo: 'bar' };
@@ -61,10 +61,10 @@ module.exports = {
       assert.equal(spy._lastRecord.message, "{ foo: 'bar' }");
     },
     'after': function() {
-      intel.console.restore();
+      rufus.console.restore();
       assert.equal(console.log, mockLog);
       console.log = prevLog;
-      intel._handlers = [];
+      rufus._handlers = [];
     }
   }