+++ test/util/error.js
@@ -2,16 +2,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const intel = require('../../');
-intel.basicConfig({
+const rufus = require('../../');
+rufus.basicConfig({
   format: '%logger.%level - %message'
 });
-intel.handleExceptions(process.argv.indexOf('--noexit') === -1);
+rufus.handleExceptions(process.argv.indexOf('--noexit') === -1);
 
 setTimeout(function() {
   throw new Error('catch me if you can');
 }, 5);
 
 setTimeout(function() {
-  intel.info('noexit');
+  rufus.info('noexit');
 }, 225);